requirements.txt
streamlit==1.38.0
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
openpyxl==3.1.5
scikit-learn==1.4.2
openai==1.52.2